+++ src/pure-function-helpers/xml-to-json.ts
@@ -1,10 +1,10 @@
 import { parseString } from 'xml2js'
-
+import { log } from '../config'
 export async function xmlToJson (xml: string): Promise<any> {
-  return new Promise((resolve, reject) => {
+  return new Promise((resolve) => {
     parseString(xml, { explicitArray: false }, (err, result) => {
       if (err) {
-        return reject(err)
+        log.warn(JSON.stringify(err))
       }
       return resolve(result)
     })
